requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
pillow==10.0.0
opencv-python==4.8.0.74
tensorflow==2.13.0 